# api.yaml
openapi: "3.0.1"
info:
  title: "OneHundredLettersApi"
  description: "API for managing correspondence, recipients and letters."
  version: "2025-03-11T16:08:24Z"
servers:
  - url: https://api.onehundredletters.com
    description: Production API
  - url: https://api-dev.onehundredletters.com
    description: Development API
security:
  - CognitoAuthorizer: []
paths:
...
  /letter/{id}:
    get:
      summary: "Get Letter By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the letter to fetch."
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "Get Letter By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    $ref: "#/components/schemas/LetterResponse"
                  message:
                    type: "string"
                    example: "Letter fetched successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    put:
      summary: "Update Letter By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the letter to update."
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LetterUpdateRequest"
        required: true
      responses:
        200:
          description: "Update Letter By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    $ref: "#/components/schemas/LetterResponse"
                  message:
                    type: "string"
                    example: "Letter updated successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Letter Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    delete:
      summary: "Delete Letter By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the letter to delete."
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "Delete Letter By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "object"
                    properties:
                      letterId:
                        $ref: "#/components/schemas/UUID"
                  message:
                    type: "string"
                    example: "Letter deleted successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
  /correspondence/{id}:
    get:
      summary: "Get Correspondence By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the correspondence to fetch."
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "Get Correspondence By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    $ref: "#/components/schemas/CorrespondenceResponse"
                  message:
                    type: "string"
                    example: "Correspondence fetched successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    put:
      summary: "Update Correspondence By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the correspondence to update."
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrespondenceUpdateRequest"
        required: true
      responses:
        200:
          description: "Update Correspondence By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    $ref: "#/components/schemas/CorrespondenceResponse"
                  message:
                    type: "string"
                    example: "Correspondence updated successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Correspondence Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    delete:
      summary: "Delete Correspondence By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the correspondence to delete."
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "Delete Correspondence By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "object"
                    properties:
                      correspondenceId:
                        $ref: '#/components/schemas/UUID'
                      recipientId:
                        $ref: '#/components/schemas/UUID'
                      letterIds:
                        type: array
                        items:
                          $ref: '#/components/schemas/UUID'
                  message:
                    type: "string"
                    example: "Correspondence deleted successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
components:
...
  schemas:
    Address:
      type: "object"
      properties:
        street:
          $ref: "#/components/schemas/Street"
        city:
          $ref: "#/components/schemas/City"
        state:
          $ref: "#/components/schemas/State"
        postalCode:
          $ref: "#/components/schemas/PostalCode"
        country:
          $ref: "#/components/schemas/Country"
      required:
        - street
        - city
        - state
        - postalCode
        - country
      example:
        street: "5623 Sunset Blvd"
        city: "Los Angeles"
        state: "CA"
        postalCode: "90047"
        country: "USA"
    Category:
      type: "string"
      enum:
        - ARTS
        - ENTERTAINMENT
        - FAMILY
        - FOOD
        - FRIENDS
        - GOVERNMENT
        - LITERATURE
        - MENTORS
        - MUSIC
        - SCIENCE
        - SPORTS
        - TECHNOLOGY
      example: "TECHNOLOGY"
    City:
      type: "string"
      example: "Los Angeles"
    CorrespondenceStatus:
      type: "string"
      enum:
        - "COMPLETED"
        - "PENDING"
        - "RESPONDED"
        - "UNSENT"
      example: "PENDING"
    CorrespondenceTitle:
      type: "string"
      example: "George Church Letter"
    Country:
      type: "string"
      example: "USA"
    DateTime:
      type: "string"
      format: "date-time"
      example: "2025-03-07T14:30:00Z"
...
    FirstName:
      type: "string"
      example: "Alan"
    FileKey:
      type: "string"
      example: "unprocessed/filename.jpg"
    LetterImage:
      type: object
      required:
        - fileKey
        - id
        - url
        - urlThumbnail
        - view
      properties:
        id:
          type: string
          example: "abc123"
        caption:
          type: string
          example: "Front of the envelope"
        dateUploaded:
          type: string
          format: date-time
          example: "2025-04-29T15:00:00Z"
        fileKey:
          $ref: "#/components/schemas/FileKey"
        mimeType:
          $ref: '#/components/schemas/MimeType'
        sizeInBytes:
          type: integer
          example: 204800
        uploadedBy:
          type: string
          example: "user@domain.com"
        url:
          type: string
          format: uri
          example: "https://example.com/letter-image1.png"
        urlThumbnail:
          type: string
          format: uri
          example: "https://example.com/letter-image1.png"
        view:
          $ref: '#/components/schemas/View'
    ImageURLs:
      type: array
      items:
        $ref: '#/components/schemas/LetterImage'
      example:
        - caption: "The front of the handwritten letter"
          dateUploaded: "2025-06-01T14:32:00Z"
          fileKey: "letters/abc123/letter_front.jpg"
          id: "9d2e3a1c-4f77-4c19-b1f0-2c0b9c713e2d"
          mimeType: "image/jpeg"
          sizeInBytes: 203845
          uploadedBy: "user_123"
          url: "https://cdn.example.com/letters/abc123/letter_front.jpg"
          urlThumbnail: "https://cdn.example.com/letters/abc123/thumb_letter_front.jpg"
          view: "LETTER_FRONT"
...
    LastName:
      type: "string"
      example: "Aldrich"
    LetterDescription:
      type: "string"
      example: "A heartfelt letter appreciating George Church's impact on high throughput DNA sequencing."
    LetterStatus:
      type: "string"
      enum:
        - "PENDING"
        - "SENT"
        - "RECEIVED"
        - "RESPONDED"
      example: "SENT"
    LetterTitle:
      type: "string"
      example: "George Church Letter"
...
    Method:
      type: "string"
      enum:
        - "TYPED"
        - "HANDWRITTEN"
        - "PRINTED"
        - "DIGITAL"
        - "OTHER"
      example: "HANDWRITTEN"
...
    Occupation:
      type: "string"
      example: "Architect"
    Organization:
      type: "string"
      example: "Sera Architects"
    PostalCode:
      type: "string"
      example: "90047"
    Reason:
      required:
        - 'description'
        - 'category'
      type: "object"
      properties:
        category:
          $ref: "#/components/schemas/Category"
        description:
          $ref: "#/components/schemas/ReasonDescription"
    ReasonDescription:
      type: "string"
      example: "Contributions to high throughput sequencing and resurrecting wooly mammoths."
    RecipientDescription:
      type: "string"
      example: "George Church's contribution to the research community is invaluable."
    State:
      type: "string"
      example: "CA"
    Street:
      type: "string"
      example: "5623 Sunset Blvd"
    Text:
      type: "string"
      example: "Dear Dr. Church, I wanted to take a moment to express my deep appreciation..."
    Type:
      type: "string"
      enum:
        - "MAIL"
        - "EMAIL"
        - "SMS"
        - "OTHER"
      example: "MAIL"
    UUID:
      type: "string"
      format: "uuid"
      example: "52b249d0-7b41-4f58-bf49-d72d592f84ff"
    View:
      type: "string"
      enum:
        - "LETTER_FRONT"
        - "LETTER_BACK"
        - "ENVELOPE_FRONT"
        - "ENVELOPE_BACK"
        - "OTHER"
      example: "ENVELOPE_BACK"
    MimeType:
      type: "string"
      enum:
        - "image/jpeg"
        - "image/png"
        - "image/webp"
        - "image/gif"
      example: "image/gif"
    LetterBase:
      type: "object"
      properties:
        correspondenceId:
          $ref: "#/components/schemas/UUID"
        description:
          $ref: "#/components/schemas/LetterDescription"
        imageURLs:
          $ref: "#/components/schemas/ImageURLs"
        method:
          $ref: "#/components/schemas/Method"
        receivedAt:
          $ref: "#/components/schemas/DateTime"
        sentAt:
          $ref: "#/components/schemas/DateTime"
        status:
          $ref: "#/components/schemas/LetterStatus"
        text:
          $ref: "#/components/schemas/Text"
        title:
          $ref: "#/components/schemas/LetterTitle"
        type:
          $ref: "#/components/schemas/Type"
    LetterResponse:
      allOf:
        - $ref: "#/components/schemas/LetterBase"
        - type: "object"
          properties:
            letterId:
              $ref: "#/components/schemas/UUID"
            createdAt:
              $ref: "#/components/schemas/DateTime"
            updatedAt:
              $ref: "#/components/schemas/DateTime"
...
    LetterUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/LetterBase"
        - type: "object"
          required:
            - "correspondenceId"
            - "imageURLs"
            - "letterId"
            - "method"
            - "status"
            - "text"
            - "title"
            - "type"
          properties:
            letterId:
              $ref: "#/components/schemas/UUID"
    RecipientBase:
      type: "object"
      properties:
        address:
          $ref: "#/components/schemas/Address"
        firstName:
          $ref: "#/components/schemas/FirstName"
        lastName:
          $ref: "#/components/schemas/LastName"
        description:
          $ref: "#/components/schemas/RecipientDescription"
        occupation:
          $ref: "#/components/schemas/Occupation"
        organization:
          $ref: "#/components/schemas/Organization"
    RecipientResponse:
      allOf:
        - $ref: "#/components/schemas/RecipientBase"
        - type: "object"
          properties:
            recipientId:
              $ref: "#/components/schemas/UUID"
            createdAt:
              $ref: "#/components/schemas/DateTime"
            updatedAt:
              $ref: "#/components/schemas/DateTime"
...
    RecipientUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/RecipientBase"
        - type: "object"
          required:
            - "address"
            - "firstName"
            - "lastName"
            - "recipientId"
          properties:
            recipientId:
              $ref: "#/components/schemas/UUID"
    Correspondence:
      type: object
      properties:
        createdAt:
          $ref: "#/components/schemas/DateTime"
        correspondenceId:
          $ref: "#/components/schemas/UUID"
        recipientId:
          $ref: "#/components/schemas/UUID"
        reason:
          $ref: "#/components/schemas/Reason"
        status: 
          $ref: "#/components/schemas/CorrespondenceStatus"
        title: 
          $ref: "#/components/schemas/CorrespondenceTitle"
        updatedAt:
          $ref: "#/components/schemas/DateTime"
...
    CorrespondenceUpdate:
      required:
        - "correspondenceId"
        - "reason"
      type: object
      properties:
        correspondenceId:
          $ref: "#/components/schemas/UUID"
        reason:
          $ref: "#/components/schemas/Reason"
        status: 
          $ref: "#/components/schemas/CorrespondenceStatus"
        title: 
          $ref: "#/components/schemas/CorrespondenceTitle"
    CorrespondenceResponse:
      type: "object"
      properties:
        correspondence:
          $ref: "#/components/schemas/Correspondence"
        recipient:
          $ref: "#/components/schemas/RecipientResponse"
        letters:
          type: array
          items:
            $ref: "#/components/schemas/LetterResponse"
...
    CorrespondenceUpdateRequest:
      required:
        - "correspondence"
        - "recipient"
        - "letters"
      type: "object"
      properties:
        correspondence:
          $ref: "#/components/schemas/CorrespondenceUpdate"
        recipient:
          $ref: "#/components/schemas/RecipientUpdateRequest"
        letters:
          type: array
          items:
            $ref: "#/components/schemas/LetterUpdateRequest"
...
    BadRequestError:
      type: "object"
      properties:
        statusCode:
          type: "integer"
          example: 400
        message:
          type: "string"
          example: "Bad Request"
    InternalServerError:
      type: "object"
      properties:
        statusCode:
          type: "integer"
          example: 500
        message:
          type: "string"
          example: "Internal Server Error"
    NotFoundError:
      type: "object"
      properties:
        statusCode:
          type: "integer"
          example: 404
        message:
          type: "string"
          example: "Not Found"
  securitySchemes:
    CognitoAuthorizer:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      description: >-
        Bearer access token. Cognito User Pool issues tokens; API Gateway validates
        them via the sibling custom TOKEN Lambda authorizer (CFN AuthorizationType CUSTOM).
        Scheme name retained for compatibility; this is not a Cognito user-pools authorizer.
      x-amazon-apigateway-authtype: "custom"